Use fixedLoadLabelAddressIntoReg for Power field watch snippets
Previously, the address of Power static and instance field watch
snippets was being loaded into a register by having the snippet manually
patch mainline code and fill in the pTOC where necessary. This has been
simplified by using fixedLoadLabelAddressIntoReg to load the address of
the snippet label. In addition to reducing code complexity, this allows
PC-relative paddi to be used when POWER10 support is enabled.

Signed-off-by: Ben Thomas <ben@benthomas.ca>