Recover gr16 as a general register in POWER10
After disableTOC is turned on in POWER10, we can recoup
gr16 as a preserved general purpose register, instead of
a reserved register. This is the first time we set up the
compiler-visible machine resources according to the CPU type.

The side-effect needs to be taken care of in JNI dispatch.
Plus, TOCBaseRegister should be prohibited from being referenced
in codegen. Otherwise, register assigner will crash.

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>