Fix class unloading processing in classBySignature cache
When caching the result of `TR_J9ServerVM::getClassFromSignature`
we incorrectly assume that the class loader of the referencing constant pool
is the same as class loader of the resulting class.
The problem is that when we delete unloaded classes from the cache,
we use the actual class loader as key, instead of referencing one
which will result in entries for unloaded classes not being deleted properly.

This commit fixes it by keeping track of which class loaders reference
a class in `ClassInfo`. When processing unloaded classes, we find
correct entries for deletion by retrieving class loaders from
`ClassInfo`.

Closes: #10397

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>