Fix max stack calculation in VT generator
Currently, instance fields are used to determine the amount of stack
space that will be used when creating a VT in the generic path. If the
VT doesn't have any instance fields then maxStack is set to zero which
is wrong. This PR ensures that maxStack is at least 1 in the generic
path.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>